+++ primitives/src/hash_types/block_hash.rs
@@ -59,12 +59,12 @@ impl encoding::Decoder for BlockHashDecoder {
 
     #[inline]
     fn push_bytes(&mut self, bytes: &mut &[u8]) -> Result<bool, Self::Error> {
-        Ok(self.0.push_bytes(bytes)?)
+        self.0.push_bytes(bytes).map_err(BlockHashDecoderError)
     }
 
     #[inline]
     fn end(self) -> Result<Self::Output, Self::Error> {
-        let a = self.0.end()?;
+        let a = self.0.end().map_err(BlockHashDecoderError)?;
         Ok(BlockHash::from_byte_array(a))
     }
 
@@ -83,10 +83,6 @@ pub struct BlockHashDecoderError(encoding::UnexpectedEofError);
 
 impl From<Infallible> for BlockHashDecoderError {
     fn from(never: Infallible) -> Self { match never {} }
-}
-
-impl From<encoding::UnexpectedEofError> for BlockHashDecoderError {
-    fn from(e: encoding::UnexpectedEofError) -> Self { Self(e) }
 }
 
 impl fmt::Display for BlockHashDecoderError {